Move multiqc_data.json export out of megaqc.py and into util_functions
